+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 0.5.17 (unreleased)
 -------------------
 
+- Add is_folder_view theme parameter : #18467
+  [laulaz]
+
 - Avoid camelcase in manifest.cfg
   See http://blog.affinitic.be/2014/08/12/beware-of-uppercase-letters-in-your-config-files/
   [laulaz]